Added txt to csv exporter
This application should read the txt file C&C Renegade produces after the end of a match, granted in its current state, this txt file needs to have the name: "resultssource.txt" otherwise it won't yet work. In the next change I will adjust it so that the user can specify which filename to read. If usernames are too long, Renegade will add # to usernames, this program should also replace this with a whitespace.